requirements.txt
# LangGraph and LangChain (aligned with al-khwarizmi-rag-guide.md)
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
langchain-anthropic>=0.2.0
langchain-chroma>=0.1.0
langchain-community>=0.3.0
langchain-huggingface>=1.2.2
langchain-openai>=1.2.1
langchain-qdrant>=1.1.0
langchain-google-genai>=4.2.2
langchain-ollama>=1.1.0

# Vector DB
chromadb>=1.2.0
qdrant-client>=1.15.0

# Embeddings (multilingual for Arabic) + optional CrossEncoder rerank
sentence-transformers>=3.0.0

# Lexical leg in hybrid retrieval (BM25 over dense candidate pool)
rank-bm25>=0.2.2

# Document loaders
pypdf>=4.0.0
docx2txt>=0.8
openpyxl>=3.1.0

# API server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
sse-starlette>=2.1.0

# Thread memory (optional; set REDIS_URL in .env)
redis>=5.0.0

# Metrics (Prometheus scrape endpoint at /metrics)
prometheus-client>=0.20.0

# Query typo normalize (pre-LLM; core/query_typo_normalize.py)
pyspellchecker>=0.8.1

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.27.0
